Extract failing test resource loader delegate
